requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
args==0.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.3
black==25.1.0
bleach==6.2.0
botocore==1.37.3
braceexpand==0.1.7
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
clint==0.5.1
clip @ https://github.com/openai/CLIP/archive/refs/heads/main.zip#sha256=11c3593912e6e6446fb0bde144c5ea374f7e19eeab9072c3eb00b59dd8afb706
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.2
cycler==0.12.1
DataProperty==1.1.0
datasets==3.3.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docutils==0.21.2
dulwich==0.22.8
einops==0.8.1
# Editable install with no version control (evaluation_library==0.0.1)
-e /home/elicer/EvalGIM
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.13
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fire==0.4.0
flake8==7.1.2
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hpsv2==1.2.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
image-reward @ git+https://github.com/THUDM/ImageReward.git@c1392c6dd0fd6ecd6d416c96959ab744a6d0a8fb
imagesize==1.4.1
importlib_metadata==8.6.1
iniconfig==2.0.0
installer==0.7.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keyring==25.6.0
kiwisolver==1.4.8
lightning-utilities==0.14.1
llvmlite==0.44.0
lxml==5.3.1
lycoris_lora==3.1.1.post1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy==1.15.0
mypy-extensions==1.0.0
myst-parser==4.0.1
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbsphinx==0.9.7
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numpy==2.2.4
numpydoc==1.8.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==2.31.0
openai==1.66.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optimizely==0.5
optimum==1.17.0
optimum-quanto==0.2.7
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paretoset==1.2.4
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
pbs-installer==2025.3.11
peft==0.9.0
pep8==1.7.1
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.12.1.2
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
recommonmark==0.7.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.23.1
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
submitit==1.5.2
sympy==1.13.1
t2v_metrics @ git+https://github.com/pietroastolfi/t2v_metrics.git@632c86585b28afff124981d80b58dffbb3538d15
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torch-fidelity==0.3.0
torchao==0.9.0
torchaudio==2.6.0+cu118
torchdiffeq==0.2.5
torcheval==0.0.7
torchmetrics==1.6.3
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.1
triton==3.2.0
trove-classifiers==2025.3.13.13
typepy==1.3.4
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.3
wandb==0.19.8
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webdataset==0.2.111
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
